Split DisplayStyle out of ViewDefinition.ts.
