+++ core/backend/src/imodeljs-backend.ts
@@ -10,6 +10,7 @@ export * from "./ClassRegistry";
 export * from "./CodeSpecs";
 export * from "./ConcurrencyControl";
 export * from "./Deprecated";
+export * from "./DisplayStyle";
 export * from "./ECDb";
 export * from "./ECSchemaXmlContext";
 export * from "./ECSqlStatement";
